Removed init files and node is in separate file
