--- Artificial_skin_PC_visualisation/connection/__init__.py
@@ -1,3 +1,0 @@
-# __init__.py
-
-from .connection import Serial
