--- Artificial_skin_PC_visualisation/debug/__init__.py
@@ -1,3 +1,0 @@
-# __init__.py
-
-from .debug import DBGLevel, debug
