--- Artificial_skin_PC_visualisation/nodes/__init__.py
@@ -1,4 +1,0 @@
-# __init__.py
-
-from .messages import prepare_bool_msg, prepare_string_msg, prepare_image_msg
-from .table import TableNode, Topic
