--- Artificial_skin_PC_visualisation/sensor/__init__.py
@@ -1,5 +1,0 @@
-# __init__.py
-
-from .sensor import Sensor
-from .params import Params, ImageMask
-from .data_parsing import cast_data_to_uint8, parse_data_to_np_image, parse_np_image_to_msg
